Recreate WordParser class
In this refactoring commit, the former WordParser class is removed and
the WordParserBase class is renamed to WordParser, which is now a non-
abstract class. The implementation of the new WordParser class is more
simpler than the previous WordParserBase class thanks to the
WordComponentParser class that was introduced in the previous commit.