Fix createTrie making dangling references
The previous implementation of Operator::createTrie was wrongly using a
initializer list of strings, in which temporary copies of the static
member constant strings are created and references to them are stored in
the operator objects in the trie.

The new implementation uses a initializer list of operator objects which
already contains valid references to the static members.