Always allocate value in SignalNumberSetImpl
A null pointer to struct sesh_osapi_sigset is not treated in the same
way as a pointer to an empty struct sesh_osapi_sigset in the
sesh_osapi_sigprocmask function. For behavioral consistency between
sesh_osapi_sigprocmask and Api::sigprocmask, SignalNumberSetImpl should
be initialized to an empty set in the constructor.