Make Jenkins test runner scripts independent
Previously, the following test runner scripts were depending on
existence of the sesh-*.tar.gz archive file:

- .jenkins-distcheck.sh
- .jenkins-doxygen.sh
- .jenkins-valgrind.sh

After this commit, those script no longer depend on the archive file.
They are now expected to be run directly in the Git workspace. It is now
easier to debug the test script in a local environment.

The .jenkins-dist.sh script is removed in favor of the new
.jenkins-archive.sh script.