Invalidate Promise before calling callback
Previously, when a result is set to a promise, the result was set to the
delay object and then the promise (delay holder) was invalidated. Which
means the corresponding callback function, if any, was being called
before the promise is invalidated. This may possibly lead to an
undefined behavior if the callback function accesses the promise. To
avoid the problem, the promise should be invalidated before the result
is set to the delay.