Invalidate Future before calling callback
Previously, when a callback function is set to a future, the callback
was set to the delay object and then the future (delay holder) was
invalidated. Which means the callback was being called before the future
is invalidated. This may possibly lead to an undefined behavior if the
callback accesses the future. To avoid the problem, the future should be
invalidated before the callback is set to the delay.