Support long file names in checkheaderinclude.sh
We have a hard limit on line length (79 characters per line), but
include guards enforced by the checkheaderinclude.sh script do not fit
in the limit if the source file name is too long. Since we are not going
to remove the limit, include guards should be truncated if too long. The
checkheaderinclude.sh script now has to accept such truncated guards.

This commit overhauls the script not only to support the long lines but
also for more user-friendly error messages.