Always declare variant copy/move constructor
This change makes the implementation simpler.

Now the copy and move constructors of variant class template
specializations are always declared as non-deleted. This means the
std::is_copy_constructible<variant<...>>::value and
std::is_move_constructible<variant<...>>::value are always true even if
the variant is actually not copy- or move-constructible. It is
apparently undesirable, but I believe it can be justified by the fact
that the same holds for std::vector<...> and many other STL class
templates.