Use tagged_union in variant
The tagged_union class template is now used to contain the variant
value, replacing the existing the value_union union.

C++11 3.8/7 requires that a manually destructed/reconstructed object be
a most derived object. With the new implementation using tagged_union,
the emplace member function family of the variant_base class template
now meets that requirement. The previous implementation was not
compliant because the variant_base object, which is a base class
subobject of the variant object, was being destructed/reconstructed
directly.