Always declare variant copy/move assignment operator
This change makes the implementation simpler.

Now the copy and move assignment operators of variant class template
specializations are always declared as non-deleted. This means the
std::is_copy_assignable<variant<...>>::value and
std::is_move_assignable<variant<...>>::value are always true even if the
variant is actually not copy- or move-assignable. It is apparently
undesirable, but I believe it can be justified by the fact that the same
holds for std::vector<...> and many other STL class templates.